+++ models/withdrawal.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 from enums.cryptocurrency import Cryptocurrency
@@ -9,8 +10,8 @@ class WithdrawalDTO(BaseModel):
     cryptoCurrency: Cryptocurrency
     toAddress: str
     txIdList: list = []
-    receivingAmount: float | None = None
-    blockchainFeeAmount: float | None = None
-    serviceFeeAmount: float | None = None
+    receivingAmount: Decimal | None = None
+    blockchainFeeAmount: Decimal | None = None
+    serviceFeeAmount: Decimal | None = None
     onlyCalculate: bool
-    totalWithdrawalAmount: float | None = None
+    totalWithdrawalAmount: Decimal | None = None
